remove lib/entrypoint.js, merge with lib/app.js (#2802)
* remove lib/app.js, merge with lib/entrypoint

* rename entrypoint.js => app.js

* fix check for built artifact
