+++ src/lib/components/base.js
@@ -1,0 +1,9 @@
+/**
+ * @fileoverview Web Components used by all pages within web.dev.
+ */
+
+import './ProfileSwitcherContainer';
+import './Header';
+import './SideNav';
+import './SnackbarContainer';
+import './Search';
